updating processor to call h41 creation

+++ app/services/report_eligibility_processor.rb
@@ -8,11 +8,11 @@ class ReportEligiblityProcessor
       if policy.present? 
         if policy.aasm_state.in?(["canceled", "carrier_canceled"])
           void_params = get_doc_params(policy, "void") 
-          transmit(void_params) if policy.federal_transmissions.present?# send VOID if current policy is canceled and there are federal transmissions present
+          transmit(void_params, policy) if policy.federal_transmissions.present?# send VOID if current policy is canceled and there are federal transmissions present
         elsif policy.aasm_state.in?(["terminated", "submitted"])
           corrected_params = get_doc_params(policy, "corrected")
           original_params = get_doc_params(policy, "original")
-          policy.federal_transmissions.present? ? transmit(corrected_params) : transmit(original_params) #send CORRECTED 1095 if a transmission is present on a submitted/termed policy, send ORGINIAL if there isn't a transmission present
+          policy.federal_transmissions.present? ? transmit(corrected_params, policy) : transmit(original_params, policy) #send CORRECTED 1095 if a transmission is present on a submitted/termed policy, send ORGINIAL if there isn't a transmission present
         end
       end
     end
@@ -46,21 +46,29 @@ class ReportEligiblityProcessor
     policy.plan.coverage_type == "health"
   end
 
-  def self.upload_to_s3(file_name, bucket_name)
-    Aws::S3Storage.save(file_name, bucket_name, File.basename(file_name), "h41")
+  def self.upload_to_s3(pfd_file, xml_file, bucket_name)
+    Aws::S3Storage.save(pfd_file, bucket_name, File.basename(file_name))
+    Aws::S3Storage.save(xml_file, bucket_name, File.basename(file_name), "h41")
   end
 
-  def self.publish_to_sftp(file_name, bucket_name)
-    Aws::S3Storage.publish_to_sftp(file_name, bucket_name, File.basename(file_name))
+  def self.publish_to_sftp(pfd_file, xml_file, bucket_name)
+    Aws::S3Storage.publish_to_sftp(pfd_file, bucket_name, File.basename(file_name))
+    Aws::S3Storage.publish_to_sftp(xml_file, bucket_name, File.basename(file_name))
   end
 
-  def self.delete_1095A_pdf(file_name)
-    File.delete(file_name)
+  def self.delete_tax_docs
+    File.delete(@pfd_file) if @pfd_file
+    File.delete(@xml_file) if @xml_file
   end 
 
   def self.generate_1095A_pdf(params)
     params[:type] = 'new' if params[:type] == 'original'
-    @file_name = Generators::Reports::IrsYearlySerializer.new(params).generate_notice
+    @pfd_file = Generators::Reports::IrsYearlySerializer.new(params).generate_notice
+  end
+  
+  def self.generate_h41_xml(params, policy)
+    #passing the true argument to trigger H41 instead of 1095A
+    @xml_file = Generators::Reports::IrsYearlySerializer.new(params, true).proccess_policy(policy)
   end
  
 
@@ -69,13 +77,13 @@ class ReportEligiblityProcessor
     federal_report.federal_report_ingester
   end
 
-  def self.transmit(params)
+  def self.transmit(params, policy)
     begin
-      if generate_1095A_pdf(params)
-         upload_to_s3(@file_name, "tax-documents")
-         publish_to_sftp(@file_name, "tax-documents")
+      if generate_1095A_pdf(params) && generate_h41_xml(params, policy)
+         upload_to_s3(@pfd_file, @xml_file, "tax-documents")
+         publish_to_sftp(@pfd_file, @xml_file, "tax-documents")
          persist_new_doc
-         delete_1095A_pdf(@file_name)
+         delete_tax_docs
          PolicyReportEligibilityUpdated.where(status: "processed").delete_all
       else
         raise("File upload failed")
